# api/openapi.yaml
openapi: 3.0.3
info:
  title: Aperture OMR
  version: "0.0.1"
  description: |
    A service to perform Optical Mark Recognition (OMR) on scanned 
    bubblesheets.

    View system resource usage [here](http://localhost:3000/dashboard.html).
servers:
  - url: http://localhost:3000/
...
paths:
...
  /system/logs:
    get:
      summary: |
        Logs
      description: |
        Check the logs from the OMR server for debugging.
      tags:
        - sysadmin
      parameters:
        - in: query
          name: limit
          required: false
          example: 10
          schema:
            type: integer
            description: |
              The number of logs to read.
      responses:
        200:
          description: |
            OK.
          content:
            text/plain:
              schema:
                description: |
                  The most recent logs from the OMR.
                type: string
        400:
          description: |
            The request was missing something. Check the error message for 
            details.
          content:
            text/plain:
              schema:
                type: string
                description: |
                  A description of the error.
  /system/utilization:
    get:
      summary: |
        Current Resource Usage
      description: |
        Check the utilization of system resources, including memory, CPU, and
        disk storage.
      tags:
        - sysadmin
      parameters:
        - in: query
          name: limit
          required: false
          example: 10
          schema:
            type: integer
            description: |
              The number of past resource utilization samples to use. There is
              one sample per second, so to get the resource utilization over 
              the past 30 seconds you would use `limit=30`.
      responses:
        200:
          description: |
            OK.
          content:
            application/json:
              schema:
                description: |
                  The most recent resource utilization for the system. The 
                  last element represents the most recent utilization sample, 
                  the second-to-last is the sample from 1 second ago, and so 
                  on.
                $ref: '#/components/schemas/ResourceUtilization'
        400:
          description: |
            The request was missing something. Check the error message for 
            details.
          content:
            text/plain:
              schema:
                type: string
                description: |
                  A description of the error.
        500:
          description: |
            There's some problem with the service. Check the response message.
          content:
            text/plain:
              schema:
                type: string
                description: |
                  A description of the error.
...
  /job:
    get:
      summary: |
        Job Status
      description: |
        Retrieves the details of a job that is currently running, or has been 
        recently completed.
      tags:
        - Job
      parameters:
        - in: query
          name: id
          required: true
          example: "00a0a849-c801-40fc-8c8b-e197ddae43cb"
          schema:
            type: string
            description: |
              The ID of the job to check.
      responses:
        200:
          description: |
            The job details were successfully retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'
        400:
          description: |
            The request was missing something. Check the error message for 
            details.
          content:
            text/plain:
              schema:
                type: string
                description: |
                  A description of the error.
        404:
          description: |
            The given `id` was not recognized. It's worth noting that completed
            jobs will be "forgotten" after a while (~24 hours).
          content:
            text/plain:
              schema:
                type: string
                description: |
                  A description of the error.
...
  /jobs:
    get:
      summary: |
        List Jobs
      description: |
        Returns all registered jobs, sorted from oldest to newest by their
        creation time.

        This endpoint requires the administrator API key.
      tags:
        - sysadmin
      parameters:
        - $ref: '#/components/parameters/AdminKeyHeader'
      responses:
        200:
          description: |
            A list of all registered jobs.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JobStatus'
        401:
          description: |
            The administrator API key was missing or invalid.
          content:
            text/plain:
              schema:
                type: string
                example: only admins can retrieve all jobs
...
components:
  parameters:
    AdminKeyHeader:
      name: OMR-Admin-Key
      in: header
      required: true
      description: Administrator authentication key.
      schema:
        type: string
  schemas:
...
    ResourceUtilization:
      type: object
      required:
        - cpuHistory
        - memoryHistory
        - memoryPeak
        - uptime
        - disk
      properties:
        cpuHistory:
          type: array
          description: Historical CPU utilization samples.
          items:
            $ref: '#/components/schemas/CpuInfo'

        memoryHistory:
          type: array
          description: Historical memory utilization samples.
          items:
            $ref: '#/components/schemas/MemInfo'
        
        memoryPeak:
          type: integer
          format: uint64
          example: 2230198
          description: |
            The peak memory usage by the OMR since startup.

        uptime:
          type: integer
          format: uint64
          example: 259219
          description: |
            The number of seconds that the OMR has been running.

        disk:
          $ref: '#/components/schemas/DiskUtilization'
    CpuInfo:
      type: object
      required:
        - description
        - overallPercent
        - mhz
        - threads
      properties:
        description:
          type: string
          description: Human-readable CPU model description.
          example: Intel(R) Core(TM) i7-12700K

        overallPercent:
          type: number
          format: double
          description: Overall CPU utilization percentage.
          example: 37.4

        mhz:
          type: number
          format: double
          description: Current operating frequency in MHz.
          example: 3600.0

        threads:
          type: array
          description: Per-thread CPU utilization.
          items:
            $ref: '#/components/schemas/CpuThreadInfo'
    CpuThreadInfo:
      type: object
      required:
        - percent
      properties:
        percent:
          type: number
          format: double
          description: CPU utilization percentage for the thread.
          example: 42.1
    MemInfo:
      type: object
      required:
        - inUseOmr
        - inUseOther
        - free
      properties:
        inUseOmr:
          type: integer
          format: uint64
          description: Bytes of memory currently used by the OMR service.
          example: 536870912

        inUseOther:
          type: integer
          format: uint64
          description: Bytes of memory used by all other processes.
          example: 7432491008

        free:
          type: integer
          format: uint64
          description: Total free system memory in bytes.
          example: 17179869184
    DiskUtilization:
      type: object
      required:
        - usage
        - omrUsage
      properties:
        usage:
          $ref: '#/components/schemas/DiskInfo'

        omrUsage:
          $ref: '#/components/schemas/OmrDiskUsage'
    DiskInfo:
      type: object
      required:
        - total
        - free
        - used
      properties:
        total:
          type: integer
          format: uint64
          description: Total disk capacity in bytes.
          example: 1000204886016

        free:
          type: integer
          format: uint64
          description: Available disk space in bytes.
          example: 642345984000

        used:
          type: integer
          format: uint64
          description: Used disk space in bytes.
          example: 357858902016
    OmrDiskUsage:
      type: object
      required:
        - database
        - numberOfMatrices
        - matrices
        - numberOfPictures
        - pictures
        - total
      properties:
        database:
          type: integer
          format: uint64
          description: Size of the OMR database in bytes.
          example: 104857600

        numberOfMatrices:
          type: integer
          description: Number of stored matrices.
          example: 245

        matrices:
          type: integer
          format: uint64
          description: Total storage used by matrices in bytes.
          example: 52428800

        numberOfPictures:
          type: integer
          description: Number of stored pictures.
          example: 1284

        pictures:
          type: integer
          format: uint64
          description: Total storage used by pictures in bytes.
          example: 734003200

        total:
          type: integer
          format: uint64
          description: Total disk usage of all OMR data in bytes.
          example: 891289600
...
    JobStatus:
      type: object
      required:
        - id
        - method
        - path
        - progress
        - startedTimestamp
        - finishedTimestamp
        - success
        - notes
      properties:
        id:
          type: string
          description: Unique identifier for the job.
          example: "00a0a849-c801-40fc-8c8b-e197ddae43cb"
        method:
          type: string
          description: HTTP method that initiated the job.
          example: POST
        path:
          type: string
          description: HTTP endpoint that initiated the job.
          example: /scan
        progress:
          type: number
          format: double
          minimum: 0
          maximum: 1
          description: |
            Fraction of the job that has completed. A value of `1.0`
            indicates the job has finished.
          example: 0.65
        startedTimestamp:
          type: integer
          format: int64
          description: |
            Unix timestamp (milliseconds since the Unix epoch) when the job
            was registered.
          example: 1753894200123
        finishedTimestamp:
          type: integer
          format: int64
          description: |
            Unix timestamp (milliseconds since the Unix epoch) when the job
            completed. A value of `-62135596800000` (the zero value of Go's
            `time.Time`) indicates the job has not yet finished.
          example: 1753894218765
        success:
          type: boolean
          description: |
            Whether the job completed successfully. This field is only
            meaningful once the job has finished.
          example: true
        notes:
          type: string
          description: |
            Optional human-readable information about the job, such as its
            size or an explanation of why it failed.
          example: "Processed 120 pages successfully."